requirements.txt
fastapi
aiogram
gunicorn
uvicorn
transformers
faiss-cpu
torch
numpy
pickle